+++ crates/miden-objects/src/batch/mod.rs
@@ -12,3 +12,6 @@ pub use proven_batch::ProvenBatch;
 
 mod proposed_batch;
 pub use proposed_batch::ProposedBatch;
+
+mod input_output_note_tracker;
+pub(crate) use input_output_note_tracker::InputOutputNoteTracker;